Добавил обработку запросов от пользователя и отправку готовых маршрутов обратно пользователю
