+++ telebot.py
@@ -1,7 +1,10 @@
 import asyncio
 from aiogram import Dispatcher, Bot, types
 from aiogram.filters import Command
+import os
 from recognition import extract_route
+from aiogram.types import FSInputFile
+import navigator
 
 
 TOKEN = '6418940022:AAGYidmVjFz8ovrFaDrfk8NrIFHeunra7k4'
@@ -14,12 +17,13 @@ async def cmd_reply(message: types.Message):
     await message.reply("Привет!\nЯ- Интеллектуальный Бот-Навигатор \nТы можешь найти маршрут от одной"
                         " точки до другой!\nВведи мне 2 точки: откуда надо построить маршрут и куда")
 
-
 @dp.message()
 async def print_route(message: types.Message):
-    text = extract_route(text=message.text)
-    print(text)
-    await message.reply(text)
+    start_room, end_room = extract_route(text=message.text).split(' ')
+    image_files = navigator.build_path(start_room, end_room)
+    for image_path in image_files:
+        photo = FSInputFile(image_path)
+        await bot.send_photo(message.chat.id, photo=photo)
 
 
 async def main():
